LayoutTests/inspector/page/setBootstrapScript-main-frame-expected.txt

Tests for Page.setBootstrapScript command.


== Running test suite: Page.setBootstrapScript
-- Running test case: Page.setBootstrapScript.MainFrame
PASS: 'valueFromBootstrapScript' should be 'undefined'.
Setting bootstrap script...
Reloading page...
PASS: 'valueFromBootstrapScript' should be '42'.
